made changes to criteria

+++ opportunity/criteria.go
@@ -8,12 +8,12 @@ import (
 type Criteria struct {
 	Id             primitive.ObjectID `bson:"_id" json:"_id"`
 	Name           string             `bson:"name" json:"name"`
-	X              float32            `bson:"x" json:"x"`
-	XII            float32            `bson:"xii" json:"xii"`
-	Cgpa           float32            `bson:"cgpa" json:"cgpa"`
+	X              float64            `bson:"x" json:"x"`
+	XII            float64            `bson:"xii" json:"xii"`
+	Cgpa           float64            `bson:"cgpa" json:"cgpa"`
 	Branches       []constant.Branch  `bson:"branches" json:"branches"`
-	ActiveBacklogs *int8              `bson:"activeBacklogs" json:"activeBacklogs"`
-	TotalBacklogs  *int8              `bson:"totalBacklogs" json:"totalBacklogs"`
+	ActiveBacklogs *int               `bson:"activeBacklogs" json:"activeBacklogs"`
+	TotalBacklogs  *int               `bson:"totalBacklogs" json:"totalBacklogs"`
 	Gender         []constant.Gender  `bson:"gender" json:"gender"`
 	Disbility      []bool             `bson:"disability" json:"disability"`
 	Courses        []constant.Course  `bson:"courses" json:"courses" validate:"required"`
